# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AqarGo API",
        "description": "AqarGo real estate platform API documentation",
        "contact": {
            "email": "support@aqargo.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8001",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/reviews": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "List all reviews (admin)",
                "operationId": "346169ce6ab0c633530a7cedc1a25fab",
                "responses": {
                    "200": {
                        "$ref": "#/components/schemas/SuccessResponse"
                    },
                    "500": {
                        "$ref": "#/components/schemas/ErrorResponse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Create a review for an ad",
                "operationId": "4c185fda015ad9617584269109afa248",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ad_id",
                                    "rating"
                                ],
                                "properties": {
                                    "ad_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "rating": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 4.5,
                                        "maximum": 5,
                                        "minimum": 0
                                    },
                                    "comment": {
                                        "type": "string",
                                        "example": "Great property!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/schemas/SuccessResponse"
                    },
                    "422": {
                        "$ref": "#/components/schemas/ValidationErrorResponse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reviews/{id}": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Get a single review",
                "operationId": "a9949d3a965ed93ae30547c07eef7505",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/schemas/SuccessResponse"
                    },
                    "404": {
                        "$ref": "#/components/schemas/ErrorResponse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Update a review",
                "operationId": "86ab7645901f456f43f875174ba762c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "rating": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 4,
                                        "maximum": 5,
                                        "minimum": 0
                                    },
                                    "comment": {
                                        "type": "string",
                                        "example": "Updated comment"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/schemas/SuccessResponse"
                    },
                    "422": {
                        "$ref": "#/components/schemas/ValidationErrorResponse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Delete a review (admin)",
                "operationId": "89986343b5dbd2b909c2764aeb7e4096",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/schemas/SuccessResponse"
                    },
                    "500": {
                        "$ref": "#/components/schemas/ErrorResponse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/client/reviews/{id}": {
            "delete": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Delete own review (client)",
                "operationId": "cf123127f1447a0dd1b7a46c5389d99f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/schemas/SuccessResponse"
                    },
                    "500": {
                        "$ref": "#/components/schemas/ErrorResponse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/ad/{ad_id}/reviews": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Get all reviews for a specific ad",
                "operationId": "12bc18cf082cbb9d41ea4484aa156283",
                "parameters": [
                    {
                        "name": "ad_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/schemas/SuccessResponse"
                    },
                    "500": {
                        "$ref": "#/components/schemas/ErrorResponse"
                    }
                }
            }
        },
        "/api/user/ad/{ad_id}/reviews": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Get the authenticated user's reviews for an ad",
                "operationId": "0870739d840931eef6e687b920dd78f4",
                "parameters": [
                    {
                        "name": "ad_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/schemas/SuccessResponse"
                    },
                    "500": {
                        "$ref": "#/components/schemas/ErrorResponse"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "SuccessResponse": {
                "properties": {
                    "status": {
                        "type": "integer",
                        "example": 1
                    },
                    "data": {
                        "type": "object"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "status": {
                        "type": "integer",
                        "example": 0
                    },
                    "data": {
                        "type": "object"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "properties": {
                    "status": {
                        "type": "integer",
                        "example": 0
                    },
                    "data": {
                        "type": "object"
                    },
                    "message": {
                        "type": "string",
                        "example": "Validation Error."
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter your JWT token"
            }
        }
    },
...
}